// web/src/shared/api/generated/model/adminSubscriptionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { BillingPeriod } from './billingPeriod';
import type { SubscriptionPlan } from './subscriptionPlan';
import type { SubscriptionStatus } from './subscriptionStatus';

export interface AdminSubscriptionResult {
  id: string;
  plan: SubscriptionPlan;
  period: BillingPeriod;
  status: SubscriptionStatus;
  currentPeriodStart: string;
  currentPeriodEnd: string;
  /** @nullable */
  entitledUntil: string | null;
  isComplimentary: boolean;
}