// web/src/shared/api/generated/model/studentProfileResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { AdminSubscriptionResult } from './adminSubscriptionResult';
import type { PlanTier } from './planTier';
import type { UserStatus } from './userStatus';

export interface StudentProfileResult {
  id: string;
  displayName: string;
  /** @nullable */
  maskedPhone: string | null;
  /** @nullable */
  maskedEmail: string | null;
  status: UserStatus;
  canSuspend: boolean;
  creationDate: string;
  /** @nullable */
  onboardedAt: string | null;
  subjectInterests: string[];
  tier: PlanTier;
  hasAskTeacher: boolean;
  subscriptions: AdminSubscriptionResult[];
}