// web/src/shared/api/generated/model/studentProgressResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { SubjectProgressResult } from './subjectProgressResult';
import type { WeakSpotsResult } from './weakSpotsResult';

export interface StudentProgressResult {
  subjects: SubjectProgressResult[];
  weakSpots: WeakSpotsResult;
}