// web/src/shared/api/generated/model/userSummaryResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PlanTier } from './planTier';
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface UserSummaryResult {
  id: string;
  displayName: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  maskedPhone: string | null;
  /** @nullable */
  maskedEmail: string | null;
  invitationPending: boolean;
  canSuspend: boolean;
  creationDate: string;
  tier: null | PlanTier;
  hasAskTeacher: boolean;
  subjectIds: string[];
}